requirements.txt
aiohttp
discord
pymongo
datetime
Pillow
